+++ src/Confluent.Kafka.OAuthBearer.Aws/AssemblyInfo.cs
@@ -1,0 +1,24 @@
+// Copyright 2026 Confluent Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using System.Runtime.CompilerServices;
+
+// Public key of test/Confluent.Kafka.OAuthBearer.Aws.UnitTests/Confluent.Kafka.OAuthBearer.Aws.UnitTests.snk
+// Strong-named IVT requires the consumer's public key because both assemblies are signed.
+[assembly: InternalsVisibleTo("Confluent.Kafka.OAuthBearer.Aws.UnitTests, " +
+                              "PublicKey=0024000004800000940000000602000000240000525341310004000001000100" +
+                              "551aa9127be4c0a77d033ceff8b4fa5132a0d914cb20d8f7cf3facc9f304c585" +
+                              "c3bb68dc61cd9833669035747c23944cd74c63efbc76e6e59217fc2144c1dab4" +
+                              "2451278a54d29ab75d112589722c45c8e6b975c8d2082e30f522fe17c2633306" +
+                              "e12bf2ba6837800817ac7f80fec1b546668a6ad839997238c3a1bf85eec9aae2")]